requirements.txt

arcgis
pandas
python-dotenv




